Add routing from login to register screen, and improve LoginScreenController
Now if the register button is clicked on the login screen, then you will be routed to the register screen. The onRegisterClicked method has also been improved to check that the two password fields match, and that the username field is not empty, and that the password field is not empty, displaying appropriate popups if these conditions are not met. Only once these conditions are met, then the register method is called, which for now only prints out the username and password, but is meant to send information to the database.