Add flexible opening hours for every day of the week for builldings.
The admin can now either select presets (All days open or weekend closed) or select custom configuration and decide the opening hours for every day separately. The modify popup doesn't have the presets. There were no changes to the database because we store the new opening hours as opening hours separated by a comma. A "Closed" option however is now possible. The WeekWidget is the object showing the opening hours and WeekDays is the object containg the actual information about the opening hours and the toString version of WeekDays is passed to the backend when adding a new building. Also changed DateBox and DayBox to public classes for everyone to use but didn't write Javadoc for them since they aren't my classes.