Make the 2 selection screens react to eachother and make the reserve button disabled if not all conditions are selected.
Makes the currently selected building of one widget disabled in the other widget. Also linked opening hours to the selected building. The reserve button will be disabled if not all the conditions are selected properly in the widgets. Wrote code for deleting a BikeReservation but not implemented anywhere in the GUI or tested yet.