Make it so that whenever a time is selected, the amount of bikes at that building will adjust according to the amount of "unclaimed" bikes.
A bike is "unclaimed" if no-one has reserved that bike for in the future. The page will first show the amount of bikes at every building after every reservation has been carried out and as soon as you click on a time, it will check in the reservations whether or not a bike has been taken from that building.